BUILD : 1.7.6 (489) BUGS : 244, 246, 247, 248, 249, 250, 251, 252, 254, 255 NOTES : 1. In some cases READONLY was not respected, and data was saved. 2. Corrected a few mistakes in example.conf. 3. Wrong column type in tables.sql for nick alias status fiag. 4. listchans and listnicks work under Cygwin. 5. NickRegDelay no longer accepts negative values. 6. -is44 option not show if converter not built. 7. Removed #ifndef STREAMLINE from the code as its no longer set during build time 8. MS MAX LIMIT was set incorrectly 9. Segfault if USERDB enabled and tables does not exist 10. Provides clear message of SUPERADMIN is not enabled
git-svn-id: svn://svn.anope.org/anope/trunk@489 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@343 5417fbe8-f217-4b02-8779-1006273d7864

+++ data/tables.sql
@@ -194,7 +194,7 @@ CREATE TABLE anope_ns_alias (
   nick varchar(255) NOT NULL default '',
   time_registered int(10) unsigned NOT NULL default '0',
   last_seen int(10) unsigned NOT NULL default '0',
-  status smallint(6) NOT NULL default '0',
+  status int(11) unsigned NOT NULL default '0',
   last_usermask text NOT NULL,
   last_realname text NOT NULL,
   last_quit text NOT NULL,
